+++ lex-web-ui/config/dist.env.js
@@ -1,8 +1,0 @@
-var merge = require('webpack-merge')
-var prodEnv = require('./prod.env')
-
-module.exports = merge(prodEnv, {
-  DIST_BUILD: '"true"',
-  PUBLIC_PATH: (process.env.PUBLIC_PATH) ?
-    `"${process.env.PUBLIC_PATH}"` : '"/"',
-})