+++ lex-web-ui/src/store/actions.js
@@ -80,6 +80,15 @@ export default {
   initConfig(context, configObj) {
     context.commit('mergeConfig', configObj);
   },
+  sendInitialUtterance(context) {
+    if (context.state.config.lex.initialUtterance) {
+      const message = {
+        type: context.state.config.ui.hideButtonMessageBubble ? 'button' : 'human',
+        text: context.state.config.lex.initialUtterance,
+      };
+      context.dispatch('postTextMessage', message);
+    }
+  },
   initMessageList(context) {
     context.commit('reloadMessages');
     if (context.state.messages && context.state.messages.length === 0) {
@@ -89,11 +98,17 @@ export default {
       });
     }
   },
-  initLexClient(context, lexRuntimeClient) {
+  initLexClient(context, payload) {
+    /* eslint-disable no-console */
+    console.log(`the payload is: ${JSON.stringify(payload, null, 2)}`);
     lexClient = new LexClient({
       botName: context.state.config.lex.botName,
       botAlias: context.state.config.lex.botAlias,
-      lexRuntimeClient,
+      lexRuntimeClient: payload.v1client,
+      botV2Id: context.state.config.lex.v2BotId,
+      botV2AliasId: context.state.config.lex.v2BotAliasId,
+      botV2LocaleId: context.state.config.lex.v2BotLocaleId,
+      lexRuntimeV2Client: payload.v2client,
     });
 
     context.commit(
@@ -478,69 +493,86 @@ export default {
         }
 
         return Promise.resolve(context.commit('pushUtterance', message.text))
-          .then(() => context.dispatch('lexPostText', message.text))
-          .then((response) => {
-            // check for an array of messages
-            if (response.message && response.message.includes('{"messages":')) {
-              const tmsg = JSON.parse(response.message);
-              if (tmsg && Array.isArray(tmsg.messages)) {
-                tmsg.messages.forEach((mes, index) => {
-                  let alts = JSON.parse(response.sessionAttributes.appContext || '{}').altMessages;
-                  if (mes.type === 'CustomPayload') {
-                    if (alts === undefined) {
-                      alts = {};
-                    }
-                    alts.markdown = mes.value;
+      })
+      .then(() => context.dispatch('lexPostText', message.text))
+      .then((response) => {
+        // check for an array of messages
+        if (response.sessionState || (response.message && response.message.includes('{"messages":'))) {
+          if (response.message && response.message.includes('{"messages":')) {
+            const tmsg = JSON.parse(response.message);
+            if (tmsg && Array.isArray(tmsg.messages)) {
+              tmsg.messages.forEach((mes, index) => {
+                let alts = JSON.parse(response.sessionAttributes.appContext || '{}').altMessages;
+                if (mes.type === 'CustomPayload' || mes.contentType === 'CustomPayload') {
+                  if (alts === undefined) {
+                    alts = {};
                   }
+                  alts.markdown = mes.value ? mes.value : mes.content;
+                }
+                // Note that Lex V1 only supported a single responseCard. V2 supports multiple response cards.
+                // This code still supports the V1 mechanism. The code below will check for
+                // the existence of a single V1 responseCard added to sessionAttributes.appContext by bots
+                // such as QnABot. This single responseCard will be appended to the last message displayed
+                // in the array of messages presented.
+                let responseCardObject = JSON.parse(response.sessionAttributes.appContext || '{}').responseCard;
+                if (responseCardObject === undefined) { // prefer appContext over lex.responseCard
+                  responseCardObject = context.state.lex.responseCard;
+                }
+                if ((mes.value && mes.value.length > 0) ||
+                  (mes.content && mes.content.length > 0)) {
                   context.dispatch(
                     'pushMessage',
                     {
-                      text: mes.value,
+                      text: mes.value ? mes.value : mes.content,
                       type: 'bot',
                       dialogState: context.state.lex.dialogState,
                       responseCard: tmsg.messages.length - 1 === index // attach response card only
-                        ? context.state.lex.responseCard : undefined, // for last response message
+                        ? responseCardObject : undefined, // for last response message
                       alts,
                     },
                   );
-                });
-              }
-            } else {
-              let alts = JSON.parse(response.sessionAttributes.appContext || '{}').altMessages;
-              let responseCardObject = JSON.parse(response.sessionAttributes.appContext || '{}').responseCard;
-              if (response.messageFormat === 'CustomPayload') {
-                if (alts === undefined) {
-                  alts = {};
                 }
-                alts.markdown = response.message;
-              }
-              if (responseCardObject === undefined) {
-                responseCardObject = context.state.lex.responseCard;
-              }
-              context.dispatch(
-                'pushMessage',
-                {
-                  text: response.message,
-                  type: 'bot',
-                  dialogState: context.state.lex.dialogState,
-                  responseCard: responseCardObject, // prefering appcontext over lex.responsecard
-                  alts,
-                },
-              );
+              });
             }
-          })
-          .then(() => {
-            if (context.state.isSFXOn) {
-              context.dispatch('playSound', context.state.config.ui.messageReceivedSFX);
+          }
+        } else {
+          let alts = JSON.parse(response.sessionAttributes.appContext || '{}').altMessages;
+          let responseCardObject = JSON.parse(response.sessionAttributes.appContext || '{}').responseCard;
+          if (response.messageFormat === 'CustomPayload') {
+            if (alts === undefined) {
+              alts = {};
             }
-            context.dispatch(
-              'sendMessageToParentWindow',
-              { event: 'messageReceived' },
-            );
-            if (context.state.lex.dialogState === 'Fulfilled') {
-              context.dispatch('reInitBot');
-            }
-          });
+            alts.markdown = response.message;
+          }
+          if (responseCardObject === undefined) {
+            responseCardObject = context.state.lex.responseCard;
+          }
+          context.dispatch(
+            'pushMessage',
+            {
+              text: response.message,
+              type: 'bot',
+              dialogState: context.state.lex.dialogState,
+              responseCard: responseCardObject, // prefering appcontext over lex.responsecard
+              alts,
+            },
+          );
+        }
+      })
+      .then(() => {
+        if (context.state.isSFXOn) {
+          context.dispatch('playSound', context.state.config.ui.messageReceivedSFX);
+          context.dispatch(
+            'sendMessageToParentWindow',
+            { event: 'messageReceived' },
+          );
+        }
+        if (context.state.lex.dialogState === 'Fulfilled') {
+          context.dispatch('reInitBot');
+        }
+        if (context.state.isPostTextRetry) {
+          context.commit('setPostTextRetry', false);
+        }
       })
       .catch((error) => {
         if (((error.message.indexOf('permissible time') === -1))