+++ src/lex-web-ui-loader/js/defaults/dependencies.js
@@ -1,5 +1,5 @@
 /*
- Copyright 2017-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ Copyright 2017-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 
  Licensed under the Amazon Software License (the "License"). You may not use this file
  except in compliance with the License. A copy of the License is located at
@@ -22,28 +22,22 @@ export const dependenciesFullPage = {
   script: [
     {
       name: 'AWS',
-      url: 'https://sdk.amazonaws.com/js/aws-sdk-2.671.0.js',
+      url: 'https://sdk.amazonaws.com/js/aws-sdk-2.875.0.js',
       canUseMin: true,
     },
     {
-      // mobile hub generated aws config
-      name: 'aws_bots_config',
-      url: './aws-config.js',
-      optional: true,
-    },
-    {
       name: 'Vue',
-      url: 'https://cdn.jsdelivr.net/npm/vue@2.5.13/dist/vue.js',
+      url: 'https://cdn.jsdelivr.net/npm/vue@2.6.11/dist/vue.js',
       canUseMin: true,
     },
     {
       name: 'Vuex',
-      url: 'https://unpkg.com/vuex@3.0.1/dist/vuex.js',
+      url: 'https://unpkg.com/vuex@3.4.0/dist/vuex.js',
       canUseMin: true,
     },
     {
       name: 'Vuetify',
-      url: 'https://unpkg.com/vuetify@0.17.6/dist/vuetify.js',
+      url: 'https://unpkg.com/vuetify@0.17.7/dist/vuetify.js',
       canUseMin: true,
     },
     {
@@ -59,7 +53,7 @@ export const dependenciesFullPage = {
     },
     {
       name: 'vuetify',
-      url: 'https://unpkg.com/vuetify@0.17.6/dist/vuetify.css',
+      url: 'https://unpkg.com/vuetify@0.17.7/dist/vuetify.css',
       canUseMin: true,
     },
     {
@@ -78,14 +72,8 @@ export const dependenciesIframe = {
   script: [
     {
       name: 'AWS',
-      url: 'https://sdk.amazonaws.com/js/aws-sdk-2.671.0.js',
+      url: 'https://sdk.amazonaws.com/js/aws-sdk-2.875.0.js',
       canUseMin: true,
-    },
-    {
-      // mobile hub generated aws config
-      name: 'aws_bots_config',
-      url: './aws-config.js',
-      optional: true,
     },
   ],
   css: [